requirements.txt
streamlit
torch
transformers
pillow
numpy
matplotlib